Remove endless loop detection to address https://github.com/linkedin/URL-Detector/issues/15.
So the obvious downside is the risk of actually getting into an endless
loop but I think the risk is worth it.  If the code is solid then this
is not possible, so I'd rather address those cases as (if?) they come
up, instead causing false failures like the case in this issue.